// browser-extension/src/client/generated/models/Recommendation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Recommendation
 */
export interface Recommendation {
    /**
     * 
     * @type {number}
     * @memberof Recommendation
     */
    tutorialId: number;
    /**
     * 
     * @type {string}
     * @memberof Recommendation
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof Recommendation
     */
    description: string;
    /**
     * 
     * @type {Date}
     * @memberof Recommendation
     */
    readonly createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Recommendation
     */
    readonly updatedAt: Date;
    /**
     * 
     * @type {string}
     * @memberof Recommendation
     */
    media: string;
    /**
     * 
     * @type {string}
     * @memberof Recommendation
     */
    link: string;
}

/**
 * Check if a given object implements the Recommendation interface.
 */
export function instanceOfRecommendation(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "tutorialId" in value;
    isInstance = isInstance && "title" in value;
    isInstance = isInstance && "description" in value;
    isInstance = isInstance && "createdAt" in value;
    isInstance = isInstance && "updatedAt" in value;
    isInstance = isInstance && "media" in value;
    isInstance = isInstance && "link" in value;

    return isInstance;
}

export function RecommendationFromJSON(json: any): Recommendation {
    return RecommendationFromJSONTyped(json, false);
}

export function RecommendationFromJSONTyped(json: any, ignoreDiscriminator: boolean): Recommendation {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'tutorialId': json['tutorial_id'],
        'title': json['title'],
        'description': json['description'],
        'createdAt': (new Date(json['created_at'])),
        'updatedAt': (new Date(json['updated_at'])),
        'media': json['media'],
        'link': json['link'],
    };
}

export function RecommendationToJSON(value?: Recommendation | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'tutorial_id': value.tutorialId,
        'title': value.title,
        'description': value.description,
        'media': value.media,
        'link': value.link,
    };
}